Remove cyclical dependencies and fix cli option addition
