+++ zipline/__init__.py
@@ -28,8 +28,7 @@ from ._version import get_versions
 # These need to happen after the other imports.
 from . algorithm import TradingAlgorithm
 from . import api
-from .extensions import Namespace, custom_types
-
+import zipline.extensions as ext
 
 # PERF: Fire a warning if calendars were instantiated during zipline import.
 # Having calendars doesn't break anything per-se, but it makes zipline imports
@@ -48,7 +47,7 @@ del global_calendar_dispatcher
 __version__ = get_versions()['version']
 del get_versions
 
-extension_args = Namespace()
+extension_args = ext.Namespace()
 
 
 def load_ipython_extension(ipython):
@@ -72,7 +71,6 @@ if os.name == 'nt':
     _()
     del _
 
-
 __all__ = [
     'TradingAlgorithm',
     'api',
@@ -82,5 +80,4 @@ __all__ = [
     'gens',
     'run_algorithm',
     'utils',
-    'custom_types',
 ]
